Moving from the static library based solution for telemetry to a header based approach to reduce complexity.
The difference (on disk) size between the lib based approach and header based approach is 1 KB (header > lib). In favor of reducing the complexity, the decision has been made to go with a header based solution
